speedo: Automatically select Authenticode signing cert.
--

This is required because GlobalSign re-issued the certificate (which
actually required to install InternetExploder in addition to Edge) and
now we have two certs to select from.  The /a option seems to use the
latest generated certificate.

(cherry picked from commit 4237a2b0a50999885c1577d27419ef8c49cd32db)