po: In German always use "Passwort" instead of "Passphrase".
--

This is a several decade old debate how to name this.  Meanwhile in
Germany it seems to be more clean to use the term "Passwort" instead
of "Passphrase" (or that "Mantra" thing).  It is easier to explain to
users that a password may include spaces etc than to to explain the
difference between passphrase and password.

So let's keep the things in the code as is but change the
translations.

Signed-off-by: Werner Koch <wk@gnupg.org>
(cherry picked from commit c9859967c0d85e36c56ff481d402b97d2fd386bb)
and adjusted for 2.3.